+++ tsconfig.json
@@ -5,6 +5,7 @@
     "allowJs": true,
     "skipLibCheck": true,
     "strict": true,
+  "baseUrl": ".",
     "noEmit": true,
     "esModuleInterop": true,
     "module": "esnext",
